+++ examples/rails_app/test_helpers/table_test_helper.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 class TableTestHelper < BaseTestHelper
-# Selectors: Semantic aliases for elements, a useful abstraction.
+# Aliases: Semantic aliases for locators, can be used in most DSL methods.
 
-# Getters: A convenient way to get related data or nested elements.
+# Finders: A convenient way to get related data or nested elements.
   def row_for(*cells)
     find(:table_row, cells)
   end