+++ test_helpers_invalid/invalid_selector_methods_test_helper.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class InvalidSelectorMethodsTestHelper < Capybara::TestHelper
-  SELECTORS = {
+  aliases(
     source: '.main',
-  }.freeze
+  )
 end