+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+*   Add `aliases` DSL to define `SELECTORS`.
+
 ## Capybara Test Helpers 1.0.1 (2020-11-19) ##
 
 *   Add `has?` alias for `has_selector?`.